debugging demo signin

+++ public/javascripts/comment.js
@@ -27,7 +27,7 @@ window.addEventListener("load", (event) => {
           })
         });
         comment.innerHTML = content;
-        gameRating.innerHTML = `Rating: ${rating}`;
+        gameRating.innerHTML = `Rating: ${rating}.0`;
         return modal.classList.toggle("show-modal");
       })
     }
@@ -35,15 +35,17 @@ window.addEventListener("load", (event) => {
     const deleteButton = document.querySelector(`.delete-comment-button.${index}`);
     const review = document.querySelector(`#review.${index}`);
 
-    deleteButton.addEventListener("click", async function(e) {
-      e.preventDefault();
+    if (deleteButton) {
+      deleteButton.addEventListener("click", async function(e) {
+        e.preventDefault();
 
-      await fetch(`/${form.action.split('/').slice(3).join('/')}`,
-        { method: 'DELETE' }
-      );
-      review.remove();
-      modal.classList.toggle("show-modal");
-    })
+        await fetch(`/${form.action.split('/').slice(3).join('/')}`,
+          { method: 'DELETE' }
+        );
+        review.remove();
+        modal.classList.toggle("show-modal");
+      })
+    }
   });
